# openapi.yml
openapi: 3.1.0
info:
  title: Jaemin Reminder API
  description: Apple Reminders Web Clone REST API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/lists:
    get:
      tags: [ReminderList]
      summary: 전체 리스트 조회
      operationId: findAllLists
      responses:
        "200":
          description: 리스트 목록 (displayOrder 순)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderListResponse"

    post:
      tags: [ReminderList]
      summary: 리스트 생성
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderListRequest"
      responses:
        "201":
          description: 생성된 리스트
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"
        "400":
          description: 유효하지 않은 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/lists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [ReminderList]
      summary: 리스트 단건 조회
      operationId: findListById
      responses:
        "200":
          description: 리스트 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [ReminderList]
      summary: 리스트 수정
      operationId: updateList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderListRequest"
      responses:
        "200":
          description: 수정된 리스트
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"
        "400":
          description: 유효하지 않은 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [ReminderList]
      summary: 리스트 삭제
      operationId: deleteList
      responses:
        "204":
          description: 삭제 완료
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/lists/reorder:
    patch:
      tags: [ReminderList]
      summary: 리스트 순서 변경
      operationId: reorderLists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
      responses:
        "204":
          description: 순서 변경 완료
        "400":
          description: 유효하지 않은 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    ReminderListRequest:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
          description: 리스트 이름
          example: 업무
        color:
          type: string
          description: HEX 색상 코드
          example: "#007AFF"
        icon:
          type: string
          description: 아이콘 이름
          example: briefcase
          nullable: true

    ReminderListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: 업무
        color:
          type: string
          example: "#007AFF"
        icon:
          type: string
          example: briefcase
          nullable: true
        displayOrder:
          type: integer
          example: 0
        createdAt:
          type: string
          format: date-time
          example: "2026-03-29T10:00:00"
        updatedAt:
          type: string
          format: date-time
          example: "2026-03-29T10:00:00"

    ReorderRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          description: 원하는 순서대로 정렬된 리스트 ID 배열
          items:
            type: integer
            format: int64
          example: [3, 1, 2]

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        message:
          type: string
          example: "ReminderList not found: 99"
        timestamp:
          type: string
          format: date-time